moved header in own component file
