+++ src/components/App/App.js
@@ -12,6 +12,7 @@ import {
 
 import { Todo } from '../TaskObjects/Todo'
 import { Add } from '../TaskObjects/Add'
+import { Header } from '../Header'
 
 const defaultTask = [
   {
@@ -37,28 +38,28 @@ function App() {
   )
 }
 
-function Header() {
-  return (
-    <>
-      <HeaderStyle>
-        <Title>SUPER IMPORTANT TODO APP</Title>
-        <NavBar>
-          <Link to="/">
-            <LinkText>Home</LinkText>
-          </Link>
+// function Header() {
+//   return (
+//     <>
+//       <HeaderStyle>
+//         <Title>SUPER IMPORTANT TODO APP</Title>
+//         <NavBar>
+//           <Link to="/">
+//             <LinkText>Home</LinkText>
+//           </Link>
 
-          <Link to="/random">
-            <LinkText>Random</LinkText>
-          </Link>
+//           <Link to="/random">
+//             <LinkText>Random</LinkText>
+//           </Link>
 
-          <Link to="/archive">
-            <LinkText>Archive</LinkText>
-          </Link>
-        </NavBar>
-      </HeaderStyle>
-    </>
-  )
-}
+//           <Link to="/archive">
+//             <LinkText>Archive</LinkText>
+//           </Link>
+//         </NavBar>
+//       </HeaderStyle>
+//     </>
+//   )
+// }
 
 function Home() {
   const [todoList, setTodo] = useState(
@@ -134,43 +135,11 @@ function Random() {
 }
 
 // TODO auslagern
-const LinkText = styled.span`
-  color: #fff;
-  textdecoration: none;
-  font-size: 1.2rem;
-  margin: 1em;
-  border: 2px solid #fff;
-  border-radius: 8px;
-  padding: 0.2em 1em;
-`
-
-const HeaderStyle = styled.header`
-  background-color: #282c34;
-  height: 10vh;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-around;
-  border-bottom: 1px solid #fff;
-`
-
-const Title = styled.h1`
-  font-size: calc(10px + 2vmin);
-  color: white;
-`
 
 const Body = styled.body`
   height: 90vh;
   background: #282c34;
   padding: 10vw;
 `
-const NavBar = styled.nav`
-  // border: 1px solid #a1a1a1;
-  border-radius: 8px;
-  width: 30vw;
-  padding: 1vh;
-  display: flex;
-  flex-flow: row;
-  justify-content: space-between;
-`
+
 export default App
